# Add performance testing: Class parsing