Unfortunately it turns out this library is quite slow - see https://github.com/xp-framework/unittest/pull/5#issuecomment-151322401

First thing to improve this it to measure performance, though.

```sh
$ xp util.profiling.Measure lang.mirrors.performance.ClassParsingPerformance -n 100
codeUnitOf(lang.XPClass<lang.mirrors.unittest.SourceTest>): 100 iteration(s), 0.103 seconds, result= true
codeUnitOf(lang.XPClass<lang.mirrors.FromReflection>): 100 iteration(s), 0.200 seconds, result= true
codeUnitOf(lang.XPClass<lang.mirrors.Sources>): 100 iteration(s), 0.002 seconds, result= true
parseDetails(lang.XPClass<lang.mirrors.unittest.SourceTest>): 100 iteration(s), 0.006 seconds, result= true
parseDetails(lang.XPClass<lang.mirrors.FromReflection>): 100 iteration(s), 0.006 seconds, result= true
parseDetails(lang.XPClass<lang.mirrors.Sources>): 100 iteration(s), 0.000 seconds, result= true
```